% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive}
\alias{archive}
\alias{git_archive_zip}
\title{Git Archive}
\usage{
git_archive_zip(file = NULL, repo = ".")
}
\arguments{
\item{file}{name of the output zip file. Default is returned
by the function}

\item{repo}{The repository to target, either as a filepath or as an object of
class \code{git_repo_ptr}. If \code{repo} is a filepath and is not a repository, then
its parent directory is considered, then the parent's parent, and so on. To
prevent this recursive search, provide a filepath protected with \code{\link[=I]{I()}}.}
}
\value{
path to the zip file that was created
}
\description{
Exports the files in your repository to a zip file that
is returned by the function.
}
\seealso{
Other git: 
\code{\link{branch}},
\code{\link{commit}},
\code{\link{diff}()},
\code{\link{fetch}},
\code{\link{git_config}()},
\code{\link{git_merge}()},
\code{\link{git_open}()},
\code{\link{repository}},
\code{\link{signature}}
}
\concept{git}
