% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{git_signature_default}
\alias{git_signature_default}
\alias{git_commit}
\title{Committing}
\usage{
git_signature_default(repo = ".")

git_commit(message, all = FALSE, repo = ".")
}
\arguments{
\item{repo}{a path to an existing repository, or a \code{git_repository} object as
returned by \link{git_open},  \link{git_init} or \link{git_clone}.}

\item{message}{a commit message}

\item{all}{if TRUE automatically adds all modified (but not new) files}
}
\description{
Write operations
}
