% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\alias{git_config}
\title{Version info}
\usage{
git_config()
}
\description{
Shows the version of libgit2 and which features have been enabled.
}
\seealso{
Other git: \code{\link{branch}}, \code{\link{commit}},
  \code{\link{git_credentials}}, \code{\link{remotes}},
  \code{\link{repository}}, \code{\link{ssh}}
}
\concept{git}
