% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential.R
\name{git_credentials}
\alias{git_credentials}
\title{Lookup passwords from the git credential store}
\usage{
git_credentials(host = "github.com", git = "git")
}
\arguments{
\item{host}{hostname to authenticate with}

\item{git}{path of the \code{git} command line program}
}
\description{
If you have the \code{git} command line program installed, this function
tries to lookup cached authentication keys from the \code{git-credential}
store.
}
\seealso{
Other git: \code{\link{branch}}, \code{\link{commit}},
  \code{\link{config}}, \code{\link{remotes}},
  \code{\link{repository}}, \code{\link{ssh}}
}
\concept{git}
