% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff}
\alias{diff}
\alias{git_diff_patch}
\title{Diff and patch}
\usage{
git_diff_patch(ref = "HEAD", parent = paste0(ref, "^"), repo = ".")
}
\arguments{
\item{ref}{string with a branch/tag/commit}

\item{parent}{the commit to compare with}

\item{repo}{The repository to target, either as a filepath or as an object of
class \code{git_repo_ptr}. If \code{repo} is a filepath and is not a repository, then
its parent directory is considered, then the parent's parent, and so on. To
prevent this recursive search, provide a filepath protected with \code{\link[=I]{I()}}.}
}
\description{
A diff is a list of changes between two trees.
}
\seealso{
Other git: 
\code{\link{branch}},
\code{\link{commit}},
\code{\link{fetch}},
\code{\link{git_config}()},
\code{\link{git_merge}()},
\code{\link{git_open}()},
\code{\link{repository}},
\code{\link{signature}}
}
\concept{git}
