% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credential.R
\name{get_git_credential}
\alias{get_git_credential}
\title{Lookup passwords from the git credential store}
\usage{
get_git_credential(host = "github.com", git = "git")
}
\arguments{
\item{host}{hostname to authenticate with}

\item{git}{path of the \code{git} command line program}
}
\description{
If you have the \code{git} command line program installed, this function
tries to lookup cached authentication keys from the \code{git-credential}
store.
}
