% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{git_archive}
\alias{git_archive}
\alias{git_archive_zip}
\title{Git Archive}
\usage{
git_archive_zip(file = NULL, repo = ".")
}
\arguments{
\item{file}{name of the output zip file. Default is returned
by the function}

\item{repo}{The path to the git repository. If the directory is not a
repository, parent directories are considered (see \code{\link[=git_find]{git_find()}}). To disable
this search, provide the filepath protected with \code{\link[=I]{I()}}. When using this
parameter, always explicitly call by name (i.e. \verb{repo = }) because future
versions of gert may have additional parameters.}
}
\value{
path to the zip file that was created
}
\description{
Exports the files in your repository to a zip file that
is returned by the function.
}
\seealso{
Other git: 
\code{\link{git_branch}()},
\code{\link{git_commit}()},
\code{\link{git_config}()},
\code{\link{git_diff}()},
\code{\link{git_fetch}()},
\code{\link{git_history}},
\code{\link{git_ignore}},
\code{\link{git_merge}()},
\code{\link{git_rebase}()},
\code{\link{git_remote}},
\code{\link{git_repo}},
\code{\link{git_reset}()},
\code{\link{git_signature}()},
\code{\link{git_stash}},
\code{\link{git_tag}}
}
\concept{git}
