% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.R
\name{git_remote}
\alias{git_remote}
\alias{git_remote_list}
\alias{git_remote_add}
\alias{git_remote_remove}
\alias{git_remote_set_url}
\alias{git_refspecs}
\title{Git Remotes}
\usage{
git_remote_list(repo = ".")

git_remote_add(name, url, refspec = NULL, repo = ".")

git_remote_remove(name, repo = ".")

git_remote_set_url(name, url, repo = ".")

git_refspecs(repo = ".")
}
\arguments{
\item{repo}{The repository to target, either as a filepath or as an object of
class \code{git_repo_ptr}. If \code{repo} is a filepath and is not a repository, then
its parent directory is considered, then the parent's parent, and so on. To
prevent this recursive search, provide a filepath protected with \code{\link[=I]{I()}}.}

\item{name}{unique name of the remote}

\item{url}{server url (https or ssh)}

\item{refspec}{optional string with the remote fetch value}
}
\description{
Add, remove and list remotes.
}
