% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodules.R
\name{git_submodule_list}
\alias{git_submodule_list}
\title{Submodules}
\usage{
git_submodule_list(repo = ".")
}
\arguments{
\item{repo}{The path to the git repository. If the directory is not a
repository, parent directories are considered (see \link{git_find}). To disable
this search, provide the filepath protected with \code{\link[=I]{I()}}. When using this
parameter, always explicitly call by name (i.e. \verb{repo = }) because future
versions of gert may have additional parameters.}
}
\description{
Interact with submodules
}
