% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{git_tag}
\alias{git_tag}
\alias{git_tag_list}
\alias{git_tag_create}
\alias{git_tag_delete}
\alias{git_tag_push}
\title{Git Tag}
\usage{
git_tag_list(match = "*", repo = ".")

git_tag_create(name, message, ref = "HEAD", repo = ".")

git_tag_delete(name, repo = ".")

git_tag_push(name, ..., repo = ".")
}
\arguments{
\item{match}{pattern to filter tags (use \code{*} for wildcard)}

\item{repo}{The repository to target, either as a filepath or as an object of
class \code{git_repo_ptr}. If \code{repo} is a filepath and is not a repository, then
its parent directory is considered, then the parent's parent, and so on. To
prevent this recursive search, provide a filepath protected with \code{\link[=I]{I()}}.}

\item{name}{tag name}

\item{message}{tag message}

\item{ref}{target reference to tag}

\item{...}{other arguments passed to \link{git_push}}
}
\description{
Create and list tags.
}
