% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{libgit2_config}
\alias{libgit2_config}
\title{Show libgit2 version and capabilities}
\usage{
libgit2_config()
}
\description{
\code{libgit2_config()} reveals which version of libgit2 gert is using and which
features are supported, such whether you are able to use ssh remotes.
}
\examples{
libgit2_config()
}
