% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.R
\name{remotes}
\alias{remotes}
\alias{git_remote_list}
\alias{git_remote_add}
\alias{git_remote_remove}
\alias{git_refspecs}
\title{Git Remotes}
\usage{
git_remote_list(repo = ".")

git_remote_add(name, url, repo = ".")

git_remote_remove(name, repo = ".")

git_refspecs(repo = ".")
}
\arguments{
\item{repo}{a path to an existing repository, or a \code{git_repository} object as
returned by \link{git_open},  \link{git_init} or \link{git_clone}.}

\item{name}{unique name of the remote}

\item{url}{server url (https or ssh)}
}
\description{
Add, remove and list remotes.
}
