% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{signature}
\alias{signature}
\alias{git_signature_default}
\alias{git_signature}
\title{Author Signature}
\usage{
git_signature_default(repo = ".")

git_signature(name, email, time = NULL)
}
\arguments{
\item{repo}{Path to the repository. If \code{path} is itself not a
repository, its parents are recursively considered, similar to
\code{\link[=git_find]{git_find()}}. An object of class \code{git_repo_ptr} is also accepted and passed
through.}

\item{name}{Real name of the committer}

\item{email}{Email address of the commmitter}

\item{time}{timestamp of class POSIXt or NULL}
}
\description{
A signature contains the author and timestamp of a commit.
This is needed by the \link{git_commit} function.
}
\seealso{
Other git: 
\code{\link{branch}},
\code{\link{commit}},
\code{\link{fetch}},
\code{\link{git_config}()},
\code{\link{git_open}()},
\code{\link{repository}}
}
\concept{git}
