% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keypair.R
\name{setup_ssh_key}
\alias{setup_ssh_key}
\title{Setup SSH Keypair}
\usage{
setup_ssh_key(file = "~/.ssh/id_rsa", open_github = TRUE)
}
\arguments{
\item{file}{destination path of the private key. For the public key, \code{.pub}
is appended to the filename.}

\item{open_github}{automatically open a browser window to let the user
add the key to Github.}
}
\description{
Generates a RSA public/private keypair if needed, and then automatically
opens the Github page to add it. Defaults to the same deafult path as the git
command line tool.
}
