% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch.R
\name{branch}
\alias{branch}
\alias{git_branch_list}
\alias{git_branch_checkout}
\alias{git_branch_create}
\alias{git_branch_delete}
\alias{git_branch_fast_forward}
\alias{git_branch_set_upstream}
\title{Git Branch}
\usage{
git_branch_list(repo = ".")

git_branch_checkout(branch, force = FALSE, repo = ".")

git_branch_create(name, ref = "HEAD", checkout = TRUE, repo = ".")

git_branch_delete(name, repo = ".")

git_branch_fast_forward(ref, repo = ".")

git_branch_set_upstream(remote = "origin", repo = ".")
}
\arguments{
\item{repo}{Path to the repository. If \code{path} is itself not a
repository, its parents are recursively considered, similar to
\code{\link[=git_find]{git_find()}}. An object of class \code{git_repo_ptr} is also accepted and passed
through.}

\item{branch}{name of branch to check out}

\item{force}{ignore conflicts and overwrite modified files}

\item{name}{string with name of the branch / tag / etc}

\item{ref}{string with a branch/tag/commit}

\item{checkout}{move HEAD to the newly created branch}

\item{remote}{name of existing remote from \link{git_remote_list}}
}
\description{
Create, list, and checkout branches.
}
\seealso{
Other git: 
\code{\link{commit}},
\code{\link{fetch}},
\code{\link{git_config}()},
\code{\link{git_open}()},
\code{\link{repository}},
\code{\link{signature}}
}
\concept{git}
