% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{commit}
\alias{commit}
\alias{git_signature_default}
\alias{git_commit}
\alias{git_add}
\alias{git_rm}
\alias{git_status}
\alias{git_ls}
\alias{git_log}
\alias{git_reset}
\title{Stage and commit changes}
\usage{
git_signature_default(repo = ".")

git_commit(message, all = FALSE, repo = ".")

git_add(files, force = FALSE, repo = ".")

git_rm(files, repo = ".")

git_status(repo = ".")

git_ls(repo = ".")

git_log(ref = "HEAD", max = 100, repo = ".")

git_reset(type = c("soft", "hard", "mixed"), ref = "HEAD",
  repo = ".")
}
\arguments{
\item{repo}{a path to an existing repository, or a \code{git_repository} object as
returned by \link{git_open},  \link{git_init} or \link{git_clone}.}

\item{message}{a commit message}

\item{all}{if TRUE automatically adds all modified (but not new) files}

\item{files}{vector of paths relative to the git root directory}

\item{force}{add files even if in gitignore}

\item{ref}{string with a branch/tag/commit}

\item{max}{lookup at most latest n parent commits}

\item{type}{must be one of \code{"soft"}, \code{"hard"}, or \code{"mixed"}}
}
\description{
To commit changes, start with \emph{staging} the files to be included in the
commit using \code{\link[=git_add]{git_add()}} (new/modified) and \code{\link[=git_rm]{git_rm()}} (deletions).
Use \code{\link[=git_status]{git_status()}} to get an overview of staged and unstaged changes, and
finally \code{\link[=git_commit]{git_commit()}} creates a new commit with currently staged files.

Also \code{\link[=git_log]{git_log()}} shows the most recent commits and \code{\link[=git_ls]{git_ls()}} lists
all the files that are being tracked in the repository.
}
\seealso{
Other git: \code{\link{branch}}, \code{\link{config}},
  \code{\link{remotes}}, \code{\link{repository}},
  \code{\link{ssh}}
}
\concept{git}
