% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff}
\alias{diff}
\alias{git_diff_list}
\alias{git_diff_patch}
\title{Git Diff}
\usage{
git_diff_list(ref = NULL, repo = ".")

git_diff_patch(ref = "HEAD", repo = ".")
}
\arguments{
\item{ref}{a reference such as \code{"HEAD"}, or a commit id, or \code{NULL}
to the diff the working directory against the repository index.}
}
\description{
View changes in a commit or in the current working directory.
}
\seealso{
Other git: 
\code{\link{archive}},
\code{\link{branch}},
\code{\link{commit}},
\code{\link{fetch}},
\code{\link{git_config}()},
\code{\link{git_merge}()},
\code{\link{git_open}()},
\code{\link{repository}},
\code{\link{signature}}
}
\concept{git}
