% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\alias{git_fetch}
\alias{git_push}
\alias{git_pull}
\title{Push and Pull}
\usage{
git_fetch(remote = NULL, refspec = NULL, password = askpass,
  ssh_key = NULL, verbose = interactive(), repo = ".")

git_push(remote = NULL, refspec = NULL, password = askpass,
  ssh_key = NULL, verbose = interactive(), repo = ".")

git_pull(..., repo = ".")
}
\arguments{
\item{remote}{name of a remote listed in \code{\link[=git_remote_list]{git_remote_list()}}}

\item{refspec}{string with mapping between remote and local refs}

\item{password}{a string or a callback function to get passwords for authentication
or password proctected ssh keys. Defaults to \link[askpass:askpass]{askpass} which
checks \code{getOption('askpass')}.}

\item{ssh_key}{path or object containing your ssh private key. By default we
look for keys in \code{ssh-agent} and \link[credentials:ssh_key_info]{credentials::ssh_key_info}.}

\item{verbose}{display some progress info while downloading}

\item{repo}{a path to an existing repository, or a \code{git_repository} object as
returned by \link{git_open},  \link{git_init} or \link{git_clone}.}

\item{...}{arguments passed to \link{git_fetch}}
}
\description{
Use \link{git_fetch} and \link{git_push} to sync a local branch with a remote.
Here \link{git_pull} is a wrapper that tries to fast-forward the local
branch after fetching.
}
\seealso{
Other git: \code{\link{branch}}, \code{\link{commit}},
  \code{\link{config}}, \code{\link{repository}}
}
\concept{git}
