% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gert-package.R
\docType{package}
\name{gert-package}
\alias{gert}
\alias{gert-package}
\title{gert: Simple Git Client for R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simple git client for R based on 'libgit2' \url{https://libgit2.org} with support for SSH and HTTPS remotes. All functions in 'gert' use basic R data types (such as vectors and data-frames) for their arguments and return values. User credentials are shared with command line 'git' through the git-credential store and ssh keys stored on disk or ssh-agent.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/gert/}
  \item \url{https://github.com/r-lib/gert}
  \item Report bugs at \url{https://github.com/r-lib/gert/issues}
}

}
\author{
\strong{Maintainer}: Jeroen Ooms \email{jeroen@berkeley.edu} (\href{https://orcid.org/0000-0002-4035-0289}{ORCID})

Other contributors:
\itemize{
  \item Jennifer Bryan \email{jenny@rstudio.com} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID}) [contributor]
}

}
\keyword{internal}
