% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{git_checkout_pull_request}
\alias{git_checkout_pull_request}
\title{GitHub Wrappers}
\usage{
git_checkout_pull_request(pr = 1, remote = "origin", repo = ".")
}
\arguments{
\item{pr}{number with PR to check out}

\item{remote}{name of a remote listed in \code{\link[=git_remote_list]{git_remote_list()}}}

\item{repo}{The path to the git repository. If the directory is not a
repository, parent directories are considered (see \link{git_find}). To disable
this search, provide the filepath protected with \code{\link[=I]{I()}}.}
}
\description{
Some wrappers for working with Github.
}
