% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{git_checkout_pull_request}
\alias{git_checkout_pull_request}
\title{GitHub Wrappers}
\usage{
git_checkout_pull_request(pr = 1, remote = "origin", repo = ".")
}
\arguments{
\item{pr}{number with PR to check out}

\item{remote}{name of a remote listed in \code{\link[=git_remote_list]{git_remote_list()}}}

\item{repo}{The repository to target, either as a filepath or as an object of
class \code{git_repo_ptr}. If \code{repo} is a filepath and is not a repository, then
its parent directory is considered, then the parent's parent, and so on. To
prevent this recursive search, provide a filepath protected with \code{\link[=I]{I()}}.}
}
\description{
Some wrappers for working with Github.
}
