% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{git_checkout_pull_request}
\alias{git_checkout_pull_request}
\title{GitHub Wrappers}
\usage{
git_checkout_pull_request(pr = 1, remote = "origin", repo = ".")
}
\arguments{
\item{pr}{number with PR to check out}

\item{remote}{name of a remote listed in \code{\link[=git_remote_list]{git_remote_list()}}}

\item{repo}{Path to the repository. If \code{path} is itself not a
repository, its parents are recursively considered, similar to
\code{\link[=git_find]{git_find()}}. An object of class \code{git_repo_ptr} is also accepted and passed
through.}
}
\description{
Some wrappers for working with Github.
}
