% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{git_config}
\alias{git_config}
\alias{git_config_global}
\alias{git_config_set}
\alias{git_config_global_set}
\alias{libgit2_config}
\title{Version info and configurations}
\usage{
git_config(repo = ".")

git_config_global()

git_config_set(name, value, repo = ".")

git_config_global_set(name, value)

libgit2_config()
}
\arguments{
\item{repo}{a path to an existing repository, or a git repository object as
returned by \code{\link[=git_open]{git_open()}}, \code{\link[=git_init]{git_init()}} or \code{\link[=git_clone]{git_clone()}}.}

\item{name}{setting name}

\item{value}{setting value, must be string, bool, number or NULL}
}
\description{
Get and set git configurations like the \code{'git config'} command line.
}
\details{
The \code{git_config} and \code{git_config_set} functions get/set options for
a given git repository, whereas \code{git_config_global} and \code{git_config_global_set}
are for global (user level) git settings, such as your username.

Use \code{libgit2_config()} to show the version of libgit2 and which features
are supported in your version of gert (such as ssh remotes).
}
\examples{
\dontrun{
# Set global git settings
git_config_global_set("user.name", "Your Name")
git_config_global_set("user.email", "your@email.com")
git_config_global()
}
# Show your libgit2 configuration:
libgit2_config()
}
\seealso{
Other git: 
\code{\link{branch}},
\code{\link{commit}},
\code{\link{fetch}},
\code{\link{repository}},
\code{\link{signature}}
}
\concept{git}
