% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{git_diff}
\alias{git_diff}
\alias{git_diff_patch}
\title{Git Diff}
\usage{
git_diff(ref = NULL, repo = ".")

git_diff_patch(ref = NULL, repo = ".")
}
\arguments{
\item{ref}{a reference such as \code{"HEAD"}, or a commit id, or \code{NULL}
to the diff the working directory against the repository index.}

\item{repo}{The repository to target, either as a filepath or as an object of
class \code{git_repo_ptr}. If \code{repo} is a filepath and is not a repository, then
its parent directory is considered, then the parent's parent, and so on. To
prevent this recursive search, provide a filepath protected with \code{\link[=I]{I()}}.}
}
\description{
View changes in a commit or in the current working directory.
}
\seealso{
Other git: 
\code{\link{branch}},
\code{\link{git_archive}},
\code{\link{git_commit}()},
\code{\link{git_config}()},
\code{\link{git_fetch}()},
\code{\link{git_merge}()},
\code{\link{git_open}()},
\code{\link{git_repo}},
\code{\link{git_signature}()}
}
\concept{git}
