% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebase.R
\name{git_rebase_info}
\alias{git_rebase_info}
\title{Cherry Pick and Rebase}
\usage{
git_rebase_info(upstream = NULL, target = "HEAD", repo = ".")
}
\arguments{
\item{upstream}{branch to which you want to rewind and re-apply your
local commits. The default uses the remote upstream branch with the
current state on the git server, simulating \link{git_pull}.}

\item{target}{branch containing the commits you want to cherry-pick
onto upstream. Defaults to current branch.}

\item{repo}{The path to the git repository. If the directory is not a
repository, parent directories are considered (see \link{git_find}). To disable
this search, provide the filepath protected with \code{\link[=I]{I()}}.}
}
\description{
Utilities for cherry-picking commits onto another branch. A rebase
operation resets the target branch to a given upstream, and then re-applies
your local changes by cherry-picking all your commits back onto the
rewinded state.
}
