% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebase.R
\name{git_rebase}
\alias{git_rebase}
\alias{git_rebase_list}
\alias{git_rebase_commit}
\alias{git_reset_hard}
\alias{git_reset_soft}
\alias{git_reset_mixed}
\alias{git_cherry_pick}
\title{Cherry-Pick and Rebase}
\usage{
git_rebase_list(upstream = NULL, repo = ".")

git_rebase_commit(upstream = NULL, repo = ".")

git_reset_hard(ref = "HEAD", repo = ".")

git_reset_soft(ref = "HEAD", repo = ".")

git_reset_mixed(ref = "HEAD", repo = ".")

git_cherry_pick(commit, repo = ".")
}
\arguments{
\item{upstream}{branch to which you want to rewind and re-apply your
local commits. The default uses the remote upstream branch with the
current state on the git server, simulating \link{git_pull}.}

\item{repo}{The path to the git repository. If the directory is not a
repository, parent directories are considered (see \link{git_find}). To disable
this search, provide the filepath protected with \code{\link[=I]{I()}}.}

\item{ref}{string with a branch/tag/commit}

\item{commit}{id of the commit to cherry pick}
}
\description{
A cherry-pick applies the changes from a given commit (from another branch)
onto the current branch. A rebase resets the branch to the state of another
branch (usually upstream) and then re-applies your local changes by cherry-
picking each of your local commits onto the upstream commit history.
}
\details{
\code{git_rebase_list} shows the commits that would need to be cherry-picked to
rebase the current branch onto \code{upstream} history. It does so by performing
a dry-run, without saving any actual changes. The output also shows if any
of the commits conflict (in which case you cannot rebase).
Use \code{git_rebase_commit} to actually rewind and rebase your local branch.
}
