% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.R
\name{git_remote}
\alias{git_remote}
\alias{git_remote_list}
\alias{git_remote_add}
\alias{git_remote_remove}
\alias{git_remote_info}
\alias{git_remote_set_url}
\alias{git_remote_set_pushurl}
\alias{git_refspecs}
\title{Git Remotes}
\usage{
git_remote_list(repo = ".")

git_remote_add(name, url, refspec = NULL, repo = ".")

git_remote_remove(name, repo = ".")

git_remote_info(name = NULL, repo = ".")

git_remote_set_url(name, url, repo = ".")

git_remote_set_pushurl(name, url, repo = ".")

git_refspecs(repo = ".")
}
\arguments{
\item{repo}{The path to the git repository. If the directory is not a
repository, parent directories are considered (see \link{git_find}). To disable
this search, provide the filepath protected with \code{\link[=I]{I()}}.}

\item{name}{unique name of the remote}

\item{url}{server url (https or ssh)}

\item{refspec}{optional string with the remote fetch value}
}
\description{
Add, remove and list remotes.
}
\seealso{
Other git: 
\code{\link{git_archive}},
\code{\link{git_branch}},
\code{\link{git_commit}()},
\code{\link{git_config}()},
\code{\link{git_diff}()},
\code{\link{git_fetch}()},
\code{\link{git_merge}()},
\code{\link{git_rebase}()},
\code{\link{git_repo}},
\code{\link{git_signature}()},
\code{\link{git_stash}},
\code{\link{git_tag}}
}
\concept{git}
