% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodules.R
\name{git_submodule_list}
\alias{git_submodule_list}
\alias{git_submodule_info}
\alias{git_submodule_init}
\alias{git_submodule_update}
\alias{git_submodule_fetch}
\alias{git_submodule_set_to}
\alias{git_submodule_add}
\title{Submodules}
\usage{
git_submodule_list(repo = ".")

git_submodule_info(submodule, repo = ".")

git_submodule_init(submodule, overwrite = FALSE, repo = ".")

git_submodule_update(submodule, init = TRUE, repo = ".")

git_submodule_fetch(submodule, ..., repo = ".")

git_submodule_set_to(submodule, ref, checkout = TRUE, repo = ".")

git_submodule_add(url, path = basename(url), ref = "HEAD", ..., repo = ".")
}
\arguments{
\item{repo}{The path to the git repository. If the directory is not a
repository, parent directories are considered (see \link{git_find}). To disable
this search, provide the filepath protected with \code{\link[=I]{I()}}. When using this
parameter, always explicitly call by name (i.e. \verb{repo = }) because future
versions of gert may have additional parameters.}

\item{submodule}{name of the submodule}

\item{overwrite}{overwrite existing entries}

\item{init}{automatically initialize before updating}

\item{...}{extra arguments for \link{git_fetch} for authentication things}

\item{ref}{a branch or tag or hash with}

\item{checkout}{actually switch the contents of the directory to this commit}

\item{url}{full git url of the submodule}

\item{path}{relative of the submodule}
}
\description{
Interact with submodules in the repository.
}
