% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{git_merge}
\alias{git_merge}
\alias{git_merge_base}
\alias{git_merge_analysis}
\alias{git_merge_stage}
\alias{git_merge_cleanup}
\title{Merging tools}
\usage{
git_merge(ref, commit_on_success = TRUE, repo = ".")

git_merge_base(ref, target = "HEAD", repo = ".")

git_merge_analysis(ref, repo = ".")

git_merge_stage(ref, repo = ".")

git_merge_cleanup(repo = ".")
}
\arguments{
\item{ref}{branch or commit that you want to merge}

\item{commit_on_success}{automatically create a merge commit if the merge succeeds without
conflicts. Set this to \code{FALSE} if you want to customize your commit message/author.}

\item{repo}{The repository to target, either as a filepath or as an object of
class \code{git_repo_ptr}. If \code{repo} is a filepath and is not a repository, then
its parent directory is considered, then the parent's parent, and so on. To
prevent this recursive search, provide a filepath protected with \code{\link[=I]{I()}}.}

\item{target}{the branch where you want to merge into. Defaults to current \code{HEAD}.}
}
\description{
Tools for preparing and performing merge operations. Under construction.
}
\seealso{
Other git: 
\code{\link{branch}},
\code{\link{commit}},
\code{\link{fetch}},
\code{\link{git_config}()},
\code{\link{git_open}()},
\code{\link{repository}},
\code{\link{signature}}
}
\concept{git}
