% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotes.R
\name{remotes}
\alias{remotes}
\alias{git_fetch}
\alias{git_push}
\alias{git_pull}
\alias{git_remotes}
\alias{git_refspecs}
\title{Git Remotes}
\usage{
git_fetch(remote = NULL, refspec = NULL, password = askpass,
  ssh_key = NULL, verbose = interactive(), repo = ".")

git_push(remote = NULL, refspec = NULL, password = askpass,
  ssh_key = NULL, verbose = interactive(), repo = ".")

git_pull(repo = ".", ...)

git_remotes(repo = ".")

git_refspecs(repo = ".")
}
\arguments{
\item{remote}{name of a remote listed in \code{\link[=git_remotes]{git_remotes()}}}

\item{refspec}{string with mapping between remote and local refs}

\item{password}{a string or a callback function to get passwords for authentication
or password proctected ssh keys. Defaults to \link[openssl:askpass]{openssl::askpass} which
checks \code{getOption('askpass')}.}

\item{ssh_key}{path or object containing your ssh private key. By default we
look for keys in \code{ssh-agent} and \link[credentials:my_ssh_key]{credentials::my_ssh_key}.}

\item{verbose}{display some progress info while downloading}

\item{repo}{a path to an existing repository, or a \code{git_repository} object as
returned by \link{git_open},  \link{git_init} or \link{git_clone}.}

\item{...}{arguments passed to \link{git_fetch}}
}
\description{
First create a repository object via \link{git_clone}, \link{git_open}, or \link{git_init}.
Then read data with \link{git_info} or \link{git_ls}.
}
\seealso{
Other git: \code{\link{branch}}, \code{\link{commit}},
  \code{\link{config}}, \code{\link{repository}}
}
\concept{git}
