% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{repository}
\alias{repository}
\alias{git_clone}
\alias{git_init}
\alias{git_open}
\alias{git_info}
\alias{git_ls}
\title{Git Repository}
\usage{
git_clone(url, path = NULL, branch = NULL, verbose = interactive())

git_init(path)

git_open(path)

git_info(repo)

git_ls(repo)
}
\arguments{
\item{url}{remote url}

\item{path}{local path, must be a non-existing or empty directory}

\item{branch}{which branch to clone}

\item{verbose}{display some progress info while downloading}

\item{repo}{a \code{git_repository} object as returned by \link{git_open},  \link{git_init} or \link{git_clone}.
If you pass a string, this will be passed to \link{git_open} first.}
}
\description{
Init, clone or open a repository.
}
\concept{git}
