% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{repository}
\alias{repository}
\alias{git_clone}
\alias{git_init}
\alias{git_open}
\alias{git_info}
\alias{git_ls}
\alias{git_add}
\alias{git_rm}
\alias{git_remotes}
\alias{git_fetch}
\title{Git Repository}
\usage{
git_clone(url, path = NULL, branch = NULL, password = askpass,
  ssh_key = my_key(), verbose = interactive())

git_init(path = ".")

git_open(path = ".")

git_info(repo = ".")

git_ls(repo = ".")

git_add(files, force = FALSE, repo = ".")

git_rm(files, repo = ".")

git_remotes(repo = ".")

git_fetch(remote = "origin", refs = NULL, repo = ".")
}
\arguments{
\item{url}{remote url. Typically starts with \code{https://github.com/} for public
repositories, and \code{https://yourname@github.com/} or \code{git@github.com/} for
private repos. You will be prompted for a password or pat when needed.}

\item{path}{local path, must be a non-existing or empty directory}

\item{branch}{which branch to clone}

\item{password}{a string or a callback function to get passwords for authentication
or password proctected ssh keys.}

\item{ssh_key}{path or object containing your ssh private key}

\item{verbose}{display some progress info while downloading}

\item{repo}{a path to an existing repository, or a \code{git_repository} object as
returned by \link{git_open},  \link{git_init} or \link{git_clone}.}

\item{files}{vector of paths relative to the git root directory}

\item{force}{add files even if in gitignore}

\item{remote}{name of a remote listed in \code{\link[=git_remotes]{git_remotes()}}}

\item{refs}{which branches/tags/etc to download. Defaults to the base refspecs.}
}
\description{
First create a repository object via \link{git_clone}, \link{git_open}, or \link{git_init}.
Then read data with \link{git_info} or \link{git_ls}.
}
\concept{git}
