% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{repository}
\alias{repository}
\alias{git_init}
\alias{git_open}
\alias{git_find}
\alias{git_info}
\title{Create or open a git repository}
\usage{
git_init(path = ".")

git_open(path = ".")

git_find(path = ".")

git_info(repo = ".")
}
\arguments{
\item{path}{local path, must be a non-existing or empty directory}

\item{repo}{a path to an existing repository, or a \code{git_repository} object as
returned by \link{git_open},  \link{git_init} or \link{git_clone}.}
}
\description{
Use \code{\link[=git_init]{git_init()}} to start a new repository or \code{\link[=git_clone]{git_clone()}} to download a
repository from a remote.
}
\seealso{
Other git: \code{\link{branch}}, \code{\link{commit}},
  \code{\link{fetch}}, \code{\link{git_config}},
  \code{\link{signature}}
}
\concept{git}
