% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{repository}
\alias{repository}
\alias{git_clone}
\alias{git_init}
\alias{git_open}
\alias{git_repository_info}
\title{Git Repository}
\usage{
git_clone(url, path = NULL, branch = NULL, verbose = FALSE)

git_init(path)

git_open(path)

git_repository_info(git_repository)
}
\arguments{
\item{url}{remote url}

\item{path}{local path, must be a non-existing or empty directory}

\item{branch}{which branch to clone}

\item{verbose}{display some progress info while downloading}

\item{git_repository}{a \code{git_repository} object as returned by \link{git_open} or
\link{git_init} or \link{git_clone}}
}
\description{
Init, clone or open a repository.
}
\concept{git}
