% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{signature}
\alias{signature}
\alias{git_signature_default}
\alias{git_signature}
\title{Author Signature}
\usage{
git_signature_default(repo = ".")

git_signature(name, email, time = NULL)
}
\arguments{
\item{repo}{a path to an existing repository, or a git repository object as
returned by \code{\link[=git_open]{git_open()}}, \code{\link[=git_init]{git_init()}} or \code{\link[=git_clone]{git_clone()}}.}

\item{name}{Real name of the committer}

\item{email}{Email address of the commmitter}

\item{time}{timestamp of class POSIXt or NULL}
}
\description{
A signature contains the author and timestamp of a commit.
This is needed by the \link{git_commit} function.
}
\seealso{
Other git: 
\code{\link{branch}},
\code{\link{commit}},
\code{\link{fetch}},
\code{\link{git_config}()},
\code{\link{repository}}
}
\concept{git}
