% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature.R
\name{signature}
\alias{signature}
\alias{git_signature_default}
\alias{git_signature}
\title{Author Signature}
\usage{
git_signature_default(repo = ".")

git_signature(name, email, time = NULL)
}
\arguments{
\item{repo}{The repository to target, either as a filepath or as an object of
class \code{git_repo_ptr}. If \code{repo} is a filepath and is not a repository, then
its parent directory is considered, then the parent's parent, and so on. To
prevent this recursive search, provide a filepath protected with \code{\link[=I]{I()}}.}

\item{name}{Real name of the committer}

\item{email}{Email address of the commmitter}

\item{time}{timestamp of class POSIXt or NULL}
}
\description{
A signature contains the author and timestamp of a commit.
This is needed by the \link{git_commit} function.
}
\seealso{
Other git: 
\code{\link{branch}},
\code{\link{commit}()},
\code{\link{fetch}()},
\code{\link{git_config}()},
\code{\link{git_open}()},
\code{\link{repository}()}
}
\concept{git}
