% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\alias{git_tag_list}
\alias{git_tag_create}
\alias{git_tag_delete}
\alias{git_tag_push}
\title{Git Tag}
\usage{
git_tag_list(match = "*", repo = ".")

git_tag_create(name, message, ref = "HEAD", repo = ".")

git_tag_delete(name, repo = ".")

git_tag_push(name, ..., repo = ".")
}
\arguments{
\item{match}{pattern to filter tags (use \code{*} for wildcard)}

\item{repo}{a path to an existing repository, or a \code{git_repository} object as
returned by \link{git_open},  \link{git_init} or \link{git_clone}.}

\item{name}{tag name}

\item{message}{tag message}

\item{ref}{target reference to tag}

\item{...}{other arguments passed to \link{git_push}}
}
\description{
Create and list tags.
}
